Store tooltip divs in state
This separates concerns between the state layer's onEachFeature and the
tooltip render code.

onEachFeature handles creating/adding/and removing the bound tooltip
div.

The tooltip element just checks if the div has been passed and, if so,
pops in the tooltip UI with a portal.

onEachFeature is wrapped in useCallback with no-deps (thanks to
functional setState) so changing spending data/category don't require
re-render of the state layer.